requirements.txt
# --- Core numerics ---
numpy
scipy
pandas
matplotlib
scikit-learn
networkx

# --- PyTorch stack ---
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
torchrl==0.10.0
tensordict==0.10.0
torch-geometric==2.6.1

# --- JAX stack ---
jax==0.6.2
jaxlib==0.6.2
optax==0.2.6
dm-haiku==0.0.15
chex==0.1.90
numpyro==0.19.0

# --- Solvers ---
gurobipy==12.0.3

# --- Misc utilities ---
dill
tqdm
gymnasium
